+++ frontend/package.json
@@ -14,6 +14,7 @@
     "date-fns": "^4.1.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-icons": "^5.5.0",
     "tailwindcss": "^4.0.7"
   },
   "devDependencies": {
